+++ core/src/main/java/app/cash/paykit/core/CashAppPay.kt
@@ -24,13 +24,15 @@ import app.cash.paykit.core.analytics.PayKitAnalyticsEventDispatcher
 import app.cash.paykit.core.analytics.PayKitAnalyticsEventDispatcherImpl
 import app.cash.paykit.core.android.ApplicationContextHolder
 import app.cash.paykit.core.exceptions.CashAppPayIntegrationException
-import app.cash.paykit.core.impl.CashAppCashAppPayImpl
+import app.cash.paykit.core.impl.CashAppPayImpl
 import app.cash.paykit.core.impl.CashAppPayLifecycleObserverImpl
 import app.cash.paykit.core.impl.NetworkManagerImpl
 import app.cash.paykit.core.models.response.CustomerResponseData
 import app.cash.paykit.core.models.sdk.CashAppPayPaymentAction
 import app.cash.paykit.core.network.OkHttpProvider
 import app.cash.paykit.core.utils.UserAgentProvider
+import app.cash.paykit.logging.CashAppLogger
+import app.cash.paykit.logging.CashAppLoggerImpl
 import kotlin.time.Duration.Companion.seconds
 
 interface CashAppPay {
@@ -135,7 +137,7 @@ object CashAppPayFactory {
 
   private val cashAppPayLifecycleObserver: CashAppPayLifecycleObserver = CashAppPayLifecycleObserverImpl()
 
-  private fun buildPayKitAnalytics(isSandbox: Boolean) =
+  private fun buildPayKitAnalytics(isSandbox: Boolean, cashAppLogger: CashAppLogger) =
     with(ApplicationContextHolder.applicationContext) {
       val info = packageManager.getPackageInfo(packageName, 0)
 
@@ -156,11 +158,12 @@ object CashAppPayFactory {
         context = ApplicationContextHolder.applicationContext,
         options = AnalyticsOptions(
           delay = 10.seconds,
-          logLevel = Log.VERBOSE,
+          logLevel = Log.WARN,
           databaseName = dbName,
           isLoggerDisabled = !BuildConfig.DEBUG,
           applicationVersionCode = versionCode!!.toInt(), // casting as int gives us the "legacy" version code
         ),
+        cashAppLogger = cashAppLogger,
       )
     }
 
@@ -180,17 +183,18 @@ object CashAppPayFactory {
       userAgentValue = getUserAgentValue(),
       okHttpClient = defaultOkHttpClient,
     )
-    val analytics = buildPayKitAnalytics(isSandbox = false)
+    val analytics = buildPayKitAnalytics(isSandbox = false, cashAppPayLogger)
     val analyticsEventDispatcher =
       buildPayKitAnalyticsEventDispatcher(clientId, networkManager, analytics, ANALYTICS_PROD_ENVIRONMENT)
     networkManager.analyticsEventDispatcher = analyticsEventDispatcher
 
-    return CashAppCashAppPayImpl(
+    return CashAppPayImpl(
       clientId = clientId,
       networkManager = networkManager,
       analyticsEventDispatcher = analyticsEventDispatcher,
       payKitLifecycleListener = cashAppPayLifecycleObserver,
       useSandboxEnvironment = false,
+      logger = cashAppPayLogger,
     )
   }
 
@@ -207,17 +211,18 @@ object CashAppPayFactory {
       okHttpClient = defaultOkHttpClient,
     )
 
-    val analytics = buildPayKitAnalytics(isSandbox = true)
+    val analytics = buildPayKitAnalytics(isSandbox = true, cashAppPayLogger)
     val analyticsEventDispatcher =
       buildPayKitAnalyticsEventDispatcher(clientId, networkManager, analytics, ANALYTICS_SANDBOX_ENVIRONMENT)
     networkManager.analyticsEventDispatcher = analyticsEventDispatcher
 
-    return CashAppCashAppPayImpl(
+    return CashAppPayImpl(
       clientId = clientId,
       networkManager = networkManager,
       analyticsEventDispatcher = analyticsEventDispatcher,
       payKitLifecycleListener = cashAppPayLifecycleObserver,
       useSandboxEnvironment = true,
+      logger = cashAppPayLogger,
     )
   }
 
@@ -240,6 +245,8 @@ object CashAppPayFactory {
   }
 
   private val defaultOkHttpClient = OkHttpProvider.provideOkHttpClient()
+
+  private val cashAppPayLogger: CashAppLogger = CashAppLoggerImpl()
 
   // Do NOT add `const` to these, as it will invalidate reflection for our Dev App.
   private val BASE_URL_SANDBOX = "https://sandbox.api.cash.app/customer-request/v1/"