+++ core/src/test/java/app/cash/paykit/core/CashAppPayAuthorizeTests.kt
@@ -17,7 +17,7 @@ package app.cash.paykit.core
 
 import app.cash.paykit.core.exceptions.CashAppPayIntegrationException
 import app.cash.paykit.core.fakes.FakeData
-import app.cash.paykit.core.impl.CashAppCashAppPayImpl
+import app.cash.paykit.core.impl.CashAppPayImpl
 import app.cash.paykit.core.models.response.CustomerResponseData
 import io.mockk.every
 import io.mockk.mockk
@@ -73,11 +73,12 @@ class CashAppPayAuthorizeTests {
   }
 
   private fun createPayKit() =
-    CashAppCashAppPayImpl(
+    CashAppPayImpl(
       clientId = FakeData.CLIENT_ID,
       networkManager = mockk(),
       payKitLifecycleListener = mockk(relaxed = true),
       useSandboxEnvironment = true,
       analyticsEventDispatcher = mockk(relaxed = true),
+      logger = mockk(relaxed = true),
     )
 }