+++ core/src/test/java/app/cash/paykit/core/NetworkRetryTests.kt
@@ -20,7 +20,7 @@ import app.cash.paykit.core.CashAppPayState.ReadyToAuthorize
 import app.cash.paykit.core.NetworkErrorTests.MockListener
 import app.cash.paykit.core.exceptions.CashAppPayConnectivityNetworkException
 import app.cash.paykit.core.fakes.FakeData
-import app.cash.paykit.core.impl.CashAppCashAppPayImpl
+import app.cash.paykit.core.impl.CashAppPayImpl
 import app.cash.paykit.core.impl.NetworkManagerImpl
 import app.cash.paykit.core.network.RetryManagerOptions
 import com.google.common.truth.Truth.assertThat
@@ -38,7 +38,9 @@ import kotlin.time.toDuration
 
 class NetworkRetryTests {
 
-  private val MAX_RETRIES = 2
+  companion object {
+    private const val MAX_RETRIES = 2
+  }
 
   @Test
   fun `failed network request will be retried and succeed`() {
@@ -188,11 +190,12 @@ class NetworkRetryTests {
   }
 
   private fun createPayKit(networkManager: NetworkManager) =
-    CashAppCashAppPayImpl(
+    CashAppPayImpl(
       clientId = FakeData.CLIENT_ID,
       networkManager = networkManager,
       payKitLifecycleListener = mockk(relaxed = true),
       useSandboxEnvironment = true,
       analyticsEventDispatcher = mockk(relaxed = true),
+      logger = mockk(relaxed = true),
     )
 }