+++ CHANGELOG.md
@@ -1,3 +1,41 @@
+# 2.3.0
+
+ - The class `CashAppPayInitializer` was made open, so that androidx.startup can be manually overridden.
+ - This version bundles fixes for minify enabled builds.
+ - Updated internal dependency on `OkHttp` to version `4.11.0`.
+ 
+## Breaking Changes
+ 
+ - Renamed class `CashAppCashAppPayApiNetworkException` to `CashAppPayApiNetworkException`
+ - Renamed class `CashAppCashAppPayApiNetworkException` to `CashAppPayApiNetworkException`
+ - This version contains a change to the bundled Cash App Pay button.
+Previously, `light` and `dark` variants of the button were made possible by using 2 different
+views, respectively `CashAppPayButtonLight` an `CashAppPayButtonDark`. As of this version, the
+there will only be a single `CashAppPayButton` view, which has been updated to support both variants.
+To obtain different variants, developers should use the XML `style` attribute to specify the variant they want, as follows:
+
+
+Light Variant:
+```xml
+<app.cash.paykit.core.ui.CashAppPayButton
+      style="@style/CAPButtonStyle.Light"
+      android:layout_height="54dp"
+      android:layout_width="match_parent"/>
+```
+
+Dark Variant:
+```xml
+<app.cash.paykit.core.ui.CashAppPayButton
+      style="@style/CAPButtonStyle.Dark"
+      android:layout_height="54dp"
+      android:layout_width="match_parent"/>
+```
+
+This change makes it possible for developer to use the button in a more flexible way, such as using
+a style that changes accordingly to the OS theme.
+
+You should migrate any instances of `CashAppPayButtonLight` and `CashAppPayButtonDark` to `CashAppPayButton`.
+
 # 2.2.1
 
 Here's what has changed on this release: